Refactor rollout completion generation to use trl.experimental.openenv.generate_rollout_completions across examples and utilities
