+++ trl/trainer/grpo_trainer.py
@@ -1160,117 +1160,6 @@ class GRPOTrainer(BaseTrainer):
         rewards_per_func = gather(rewards_per_func)
         return rewards_per_func
 
-    def _build_colocate_sampling_params(
-        self,
-        overrides: dict[str, Any] | None = None,
-        *,
-        logprobs: bool = True,
-    ) -> "SamplingParams":
-        if self.guided_decoding_regex:
-            guided_decoding = GuidedDecodingParams(regex=self.guided_decoding_regex)
-        else:
-            guided_decoding = None
-
-        generation_kwargs: dict[str, Any] = {
-            "n": 1,
-            "temperature": self.temperature,
-            "top_k": -1 if self.top_k is None else self.top_k,
-            "min_p": 0.0 if self.min_p is None else self.min_p,
-            "max_tokens": self.max_completion_length,
-            "truncate_prompt_tokens": self.max_prompt_length,
-            "guided_decoding": guided_decoding,
-        }
-        if self.repetition_penalty is not None:
-            generation_kwargs["repetition_penalty"] = self.repetition_penalty
-        if self.top_p is not None:
-            generation_kwargs["top_p"] = self.top_p
-        if logprobs:
-            generation_kwargs["logprobs"] = 0
-
-        if self.args.generation_kwargs is not None:
-            generation_kwargs.update(self.args.generation_kwargs)
-
-        if overrides is not None:
-            generation_kwargs.update(overrides)
-
-        generation_kwargs = {key: value for key, value in generation_kwargs.items() if value is not None}
-
-        sampling_params = SamplingParams(**generation_kwargs)
-        if sampling_params.n != 1:
-            raise ValueError("generate_rollout_completions expects n=1 when using colocated vLLM.")
-        return sampling_params
-
-    def generate_rollout_completions(
-        self,
-        prompts: list[str],
-        *,
-        generation_overrides: dict[str, Any] | None = None,
-        as_chat: bool | None = None,
-    ) -> list[dict[str, Any]]:
-        """
-        Generate completions for custom rollouts when vLLM is running in colocate mode.
-
-        Returns one result per prompt, containing prompt and completion token ids along with per-token log
-        probabilities and the generated text.
-        """
-
-        if not prompts:
-            return []
-
-        if not self.use_vllm or self.vllm_mode != "colocate":
-            raise RuntimeError("Custom rollouts require vLLM in colocate mode to call generate_rollout_completions.")
-
-        sampling_params = self._build_colocate_sampling_params(generation_overrides)
-        prompts_for_generation = prompts
-        original_size = len(prompts)
-
-        if self.vllm_tensor_parallel_size > 1:
-            gathered_prompts = [None for _ in range(self.vllm_tensor_parallel_size)]
-            torch.distributed.all_gather_object(gathered_prompts, prompts, group=self.tp_group)
-            prompts_for_generation = [prompt for group_prompts in gathered_prompts for prompt in group_prompts]
-
-        if as_chat is None:
-            as_chat = prompts_for_generation and is_conversational({"prompt": prompts_for_generation[0]})
-
-        if self.args.vllm_enable_sleep_mode:
-            self.llm.wake_up(tags=["kv_cache"])
-
-        with profiling_context(self, "vLLM.generate_rollout"):
-            if as_chat:
-                vllm_outputs = self.llm.chat(prompts_for_generation, sampling_params=sampling_params, use_tqdm=False)
-            else:
-                vllm_outputs = self.llm.generate(
-                    prompts_for_generation, sampling_params=sampling_params, use_tqdm=False
-                )
-
-        results: list[dict[str, Any]] = []
-        for request in vllm_outputs:
-            if not request.outputs:
-                results.append(
-                    {"prompt_ids": request.prompt_token_ids, "completion_ids": [], "logprobs": [], "text": ""}
-                )
-                continue
-            sequence = request.outputs[0]
-            logprobs = [next(iter(token_logprob.values())).logprob for token_logprob in sequence.logprobs]
-            results.append(
-                {
-                    "prompt_ids": request.prompt_token_ids,
-                    "completion_ids": sequence.token_ids,
-                    "logprobs": logprobs,
-                    "text": sequence.text,
-                }
-            )
-
-        if self.vllm_tensor_parallel_size > 1:
-            local_rank_in_group = torch.distributed.get_rank(group=self.tp_group)
-            tp_slice = slice(local_rank_in_group * original_size, (local_rank_in_group + 1) * original_size)
-            results = results[tp_slice]
-
-        if self.args.vllm_enable_sleep_mode:
-            self.llm.sleep(level=2)
-
-        return results
-
     def _generate_single_turn(self, prompts: list):
         device = self.accelerator.device
